+++ packages/rich-text/src/component/index.native.js
@@ -4,7 +4,7 @@
  * External dependencies
  */
 import { View, Platform } from 'react-native';
-import { get, pickBy, debounce, isString } from 'lodash';
+import { get, pickBy, debounce } from 'lodash';
 import memize from 'memize';
 
 /**
@@ -56,6 +56,7 @@ const gutenbergFormatNamesToAztec = {
 };
 
 const EMPTY_PARAGRAPH_TAGS = '<p></p>';
+const DEFAULT_FONT_SIZE = 16;
 
 export class RichText extends Component {
 	constructor( {
@@ -111,9 +112,6 @@ export class RichText extends Component {
 		).bind( this );
 		this.suggestionOptions = this.suggestionOptions.bind( this );
 		this.insertString = this.insertString.bind( this );
-		this.convertFontSizeFromString = this.convertFontSizeFromString.bind(
-			this
-		);
 		this.manipulateEventCounterToForceNativeToRefresh = this.manipulateEventCounterToForceNativeToRefresh.bind(
 			this
 		);
@@ -275,13 +273,6 @@ export class RichText extends Component {
 		return html
 			.replace( openingTagRegexp, '' )
 			.replace( closingTagRegexp, '' );
-	}
-
-	// Fix for crash https://github.com/wordpress-mobile/gutenberg-mobile/issues/2991
-	convertFontSizeFromString( fontSize ) {
-		return fontSize && isString( fontSize ) && fontSize.endsWith( 'px' )
-			? parseFloat( fontSize.substring( 0, fontSize.length - 2 ) )
-			: fontSize;
 	}
 
 	/*
@@ -764,6 +755,14 @@ export class RichText extends Component {
 				this.needsSelectionUpdate = true;
 				this.manipulateEventCounterToForceNativeToRefresh(); // force a refresh on the native side
 			}
+
+			if (
+				nextProps?.style?.fontSize !== this.props?.style?.fontSize ||
+				nextProps?.style?.lineHeight !== this.props?.style?.lineHeight
+			) {
+				this.needsSelectionUpdate = true;
+				this.manipulateEventCounterToForceNativeToRefresh(); // force a refresh on the native side
+			}
 		}
 
 		return true;
@@ -853,6 +852,44 @@ export class RichText extends Component {
 		};
 	}
 
+	getFontSize() {
+		const { baseGlobalStyles } = this.props;
+
+		if ( this.props.fontSize ) {
+			return parseFloat( this.props.fontSize );
+		}
+
+		if ( this.props.style?.fontSize ) {
+			return parseFloat( this.props.style.fontSize );
+		}
+
+		if ( baseGlobalStyles?.typography?.fontSize ) {
+			return parseFloat( baseGlobalStyles?.typography?.fontSize );
+		}
+
+		return DEFAULT_FONT_SIZE;
+	}
+
+	getLineHeight() {
+		const { baseGlobalStyles } = this.props;
+		let lineHeight;
+
+		// eslint-disable-next-line no-undef
+		if ( ! __DEV__ ) {
+			return;
+		}
+
+		if ( baseGlobalStyles?.typography?.lineHeight ) {
+			lineHeight = parseFloat( baseGlobalStyles?.typography?.lineHeight );
+		}
+
+		if ( this.props.style?.lineHeight ) {
+			lineHeight = parseFloat( this.props.style.lineHeight );
+		}
+
+		return lineHeight;
+	}
+
 	render() {
 		const {
 			tagName,
@@ -879,6 +916,8 @@ export class RichText extends Component {
 		);
 
 		const { color: defaultPlaceholderTextColor } = placeholderStyle;
+		const fontSize = this.getFontSize();
+		const lineHeight = this.getLineHeight();
 
 		const {
 			color: defaultColor,
@@ -1017,11 +1056,8 @@ export class RichText extends Component {
 					}
 					maxImagesWidth={ 200 }
 					fontFamily={ this.props.fontFamily || defaultFontFamily }
-					fontSize={
-						this.props.fontSize ||
-						( style &&
-							this.convertFontSizeFromString( style.fontSize ) )
-					}
+					fontSize={ fontSize }
+					lineHeight={ lineHeight }
 					fontWeight={ this.props.fontWeight }
 					fontStyle={ this.props.fontStyle }
 					disableEditingMenu={ disableEditingMenu }
@@ -1079,8 +1115,9 @@ export default compose( [
 			'attributes',
 			'childrenStyles',
 		] );
-		const baseGlobalStyles = getSettings()
-			?.__experimentalGlobalStylesBaseStyles;
+
+		const settings = getSettings();
+		const baseGlobalStyles = settings?.__experimentalGlobalStylesBaseStyles;
 
 		return {
 			areMentionsSupported: